add the g:indentLine_setColors option
When g:indentLine_setColors is 1, indentLine will set the Conceal
highlight group's colors using the g:indentLine_color_{term,gui} options
if provided, or choose default colors automatically if not (this is the
existing behavior.) When g:indentLine_setColors is 0, indentLine will
not override the existing colors selected for the Conceal highlight
group. Defaults to 1.